# openapi-draft.yaml
openapi: 3.1.0
info:
  title: AI OS Conversational Agent Builder API
  version: 0.1.0-draft
  description: >-
    Architecture-level API contract for teaching, validating, approving, publishing and running
    enterprise Agents. Langflow is hidden behind an Agent Runtime Adapter and is not the system of record.
servers:
  - url: https://aios.example.com
security:
  - bearerAuth: []
...
paths:
...
  /v1/runs/{runId}:
    get:
      tags: [Runs]
      summary: Get run status and summary
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Run'}}}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RunId: {name: runId, in: path, required: true, schema: {type: string}}
  schemas:
...
    Run:
      type: object
      properties:
        id: {type: string}
        status: {enum: [queued, running, suspended, succeeded, failed, cancelled]}
        agentVersion: {type: string}
        runtime: {type: string}
        startedAt: {type: string, format: date-time}
        completedAt: {type: string, format: date-time}
        result: {type: object}
        cost: {type: object}
        traceId: {type: string}